modified to move auth class
